[ MVP ] Rename BasePresenter.onViewReady() to BasePresenter.start().
Start presenter business when its related view is ready.
Presenter doesn't make sense without a view, and only has a 1-to-1 relation with the view.  Presenter maybe need to do some initialization stuff when the view is ready.  That's the reason we define this method which should be invoked when the view is ready, for example, when Fragment.onViewCreated, Activity.onResume.