[ Kotlin+MVVM ]  Fix build error "Entities and Pojos must have a usable public constructor" by updating Room from 1.0.0 to 1.1.1.
florina-muntenescu commented on 14 Oct 2017
Thanks @rs146 for reporting this. The problem here is that Room can't find the right constructor for the elements in the relation, when the element is Int (or Integer in Java). I created an issue on our internal tracker for this but chances are this won't be fixed for the 1.0 release, unfortunately.
For now, the workaround is to return List<PersonSkill> instead of List<Int> for the element annotated with @Relation

https://github.com/googlesamples/android-architecture/issues/434